pyproject.toml
[project]
name = "python-template-for-ai-assistant"
dynamic = ["version"]
description = "A modern Python project template optimized for Claude Code with strict guardrails"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.ruff]
include = ["src/**/*.py", "tests/**/*.py", "manual_tests/**/*.py"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "PyYAML>=6.0.2",
    "ruff>=0.4.0",
    "ty==0.0.24",
    "taskipy>=1.14.1",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=2.0.0",
    "myst-parser>=2.0.0",
]


# =====================
# taskipy タスク定義
# =====================
[tool.taskipy.tasks]
check = """
  uv run python scripts/git_hook_runner.py pre-commit --all-files &&
  uv run pytest tests
"""
